refactor(additional-actions): move to root
